requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==4.1.0
djangorestframework==3.14.0
Faker==18.11.2
gunicorn==19.9.0
Pillow==9.5.0
psycopg2==2.9.3
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.0
